+++ src/main/java/com/beanSpot/WEB3_4_Poten_BE/domain/cafe/controller/CafeController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.dto.CafeInfoResponse;
-import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.dto.CafeUpdateRequest;
+import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.dto.res.CafeInfoRes;
+import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.dto.req.CafeUpdateReq;
 import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.entity.Cafe;
 import com.beanSpot.WEB3_4_Poten_BE.domain.cafe.service.CafeService;
 
@@ -24,13 +24,13 @@ public class CafeController {
 	private final CafeService cafeService;
 
 	@PutMapping("/{id}")
-	public ResponseEntity<Cafe> updateCafe(@PathVariable Long id, @RequestBody CafeUpdateRequest request) {
+	public ResponseEntity<Cafe> updateCafe(@PathVariable Long id, @RequestBody CafeUpdateReq request) {
 		Cafe updatedCafe = cafeService.updateCafe(id, request);
 		return ResponseEntity.ok(updatedCafe);
 	}
 
 	@GetMapping
-	public List<CafeInfoResponse> getCafeList() {
+	public List<CafeInfoRes> getCafeList() {
 		return cafeService.getCafeList();
 	}
 }
