+++ src/main/java/com/beanSpot/WEB3_4_Poten_BE/domain/cafe/dto/req/CafeUpdateReq.java
@@ -1,6 +1,6 @@
-package com.beanSpot.WEB3_4_Poten_BE.domain.cafe.dto;
+package com.beanSpot.WEB3_4_Poten_BE.domain.cafe.dto.req;
 
-public record CafeUpdateRequest(
+public record CafeUpdateReq(
 	String name,
 	String address,
 	String phone,
